+++ use_cases/config.py
@@ -1,6 +1,10 @@
 from adalflow.components.model_client.groq_client import GroqAPIClient
 from adalflow.components.model_client.openai_client import OpenAIClient
 
+from adalflow.utils import setup_env
+
+setup_env()
+
 
 llama3_model = {
     "model_client": GroqAPIClient(),
@@ -21,6 +25,19 @@ gpt_3_model = {
     },
 }
 
+# https://openai.com/api/pricing/
+# use this for evaluation
+gpt_4o_mini_model = {
+    "model_client": OpenAIClient(),
+    "model_kwargs": {
+        "model": "gpt-4o-mini",
+        "temperature": 1,
+        "top_p": 0.99,
+        "max_tokens": 1000,
+        # "frequency_penalty": 1,  # high for nto repeating prompt
+    },
+}
+
 gpt_4o_model = {
     "model_client": OpenAIClient(),
     "model_kwargs": {
@@ -31,5 +48,6 @@ gpt_4o_model = {
         # "frequency_penalty": 1,  # high for nto repeating prompt
     },
 }
+
 
 dataset_path = "cache_datasets"